+++ lib/racecar.rb
@@ -10,7 +10,7 @@ module Racecar
   # Ignores all instrumentation events.
   class NullInstrumenter
     def self.instrument(*)
-      yield if block_given?
+      yield({}) if block_given?
     end
   end