+++ lib/racecar/runner.rb
@@ -29,8 +29,9 @@ module Racecar
       # Configure the consumer with a producer so it can produce messages and
       # with a consumer so that it can support advanced use-cases.
       processor.configure(
-        producer: producer,
-        consumer: consumer
+        producer:     producer,
+        consumer:     consumer,
+        instrumenter: @instrumenter,
       )
 
       # Main loop
@@ -66,7 +67,9 @@ module Racecar
 
     def producer
       # https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md
-      @producer ||= Rdkafka::Config.new(producer_config).producer
+      @producer ||= Rdkafka::Config.new(producer_config).producer.tap do |producer|
+        producer.delivery_callback = delivery_callback
+      end
     end
 
     def producer_config
@@ -77,6 +80,13 @@ module Racecar
       producer_config["compression.codec"] = config.producer_compression_codec.to_s unless config.producer_compression_codec.nil?
       producer_config.merge(config.rdkafka_producer)
       producer_config
+    end
+
+    def delivery_callback
+      ->(delivery_report) do
+        data = {offset: delivery_report.offset, partition: delivery_report.partition}
+        @instrumenter.instrument("acknowledged_message.racecar", data)
+      end
     end
 
     def install_signal_handlers